requirements.txt
gunicorn==20.1.0
Flask==2.2.5
python-dotenv==1.0.0
groq
langchain-groq
flask-cors
sqlalchemy
psycopg2-binary
redis
